Adding cmd line parsing, list operations and store mutated code options

+++ bugbane.py
@@ -1,4 +1,5 @@
 import ast
+import argparse
 import copy
 import inspect
 import sys
@@ -139,14 +140,26 @@ def replace_import_statement(original_file ,test_file, mutant_file,reverse = Fal
                     
             f.write(line)
 
-def run_bugbane():
-    if len(sys.argv) < 3:
-        print("Usage: python mutant_tester.py <original_file> <test_file>")
-        return
+def build_parser():
+    parser = argparse.ArgumentParser(description='Bugbane - A comprehensive Mutation Testing Tool for Python Source Code',
+                                     fromfile_prefix_chars='@')
+    
+    parser.add_argument('--source-file', '-s', type=str, nargs='+', help='target module or package to mutate')
+    parser.add_argument('--test-file', '-t', type=str, nargs='+',
+                        help='test class, test method, module or package with unit tests for the target')
+    
+    parser.add_argument('--show-mutants', '-m', action='store_true', help='show all mutant source codes')
+    
+    parser.add_argument('--list-operators', '-l', action='store_true', help='list available mutation operators')
+    return parser
 
-    original_filename = sys.argv[1]
-    test_filename = sys.argv[2]
+def list_mutation_operators(mutation_operators):
+    for mutation_operator in mutation_operators:
+        print(type(mutation_operator).__name__)
+    
 
+
+def run_bugbane(parser):
     mutation_operators = [
         ArithmeticOperatorMutationOperator([ast.Add, ast.Sub, ast.Mult, ast.Div]),
         LogicalOperatorMutationOperator([ast.And, ast.Or, ast.Not]),
@@ -163,32 +176,49 @@ def run_bugbane():
         NegateConditionalsMutator(),
         EmptyReturnsMutator()
     ]
+    config = parser.parse_args()
 
-    mutants = apply_mutations_to_file(original_filename, mutation_operators)
-    test_mutants = generate_mutant_test_files(original_filename,test_filename,mutants)
-    number_of_mutants = len(mutants)
-    number_of_test_failed = 0
-    number_of_test_passed = 0
-    for index in range(number_of_mutants):  
-        if run_file_against_tests(mutants[index],test_mutants[index],original_filename):
-            number_of_test_passed += 1
-        else:
-            number_of_test_failed += 1
-    print("Number of Mutants Passed: ",number_of_test_passed)
-    print("Number of Mutants Failed: ",number_of_test_failed)
-    print("Mutation Score: ", (number_of_test_failed/number_of_mutants)*100)
+    if config.list_operators:
+        list_mutation_operators(mutation_operators)
+    
+    # if len(sys.argv) < 3:
+    #     print("Usage: python mutant_tester.py <original_file> <test_file>")
+    #     return
 
-    # Remove mutated files and modified test files
-    for mutant_filename in mutants:
-        os.remove(mutant_filename)
-        modified_test_filename = f"{test_filename}-mutant"
-        if os.path.exists(modified_test_filename):
-            os.remove(modified_test_filename)
-    for mutant_test_filename in test_mutants:
-        os.remove(mutant_test_filename)
-        modified_test_filename = f"{test_filename}-mutant-test"
-        if os.path.exists(modified_test_filename):
-            os.remove(modified_test_filename)
+    # original_filename = sys.argv[1]
+    # test_filename = sys.argv[2]
+
+    
+    if config.source_file and config.test_file:
+        original_filename = config.source_file[0]
+        test_filename = config.test_file[0]
+        print(original_filename,test_filename)
+        mutants = apply_mutations_to_file(original_filename, mutation_operators)
+        test_mutants = generate_mutant_test_files(original_filename,test_filename,mutants)
+        number_of_mutants = len(mutants)
+        number_of_test_failed = 0
+        number_of_test_passed = 0
+        for index in range(number_of_mutants):  
+            if run_file_against_tests(mutants[index],test_mutants[index],original_filename):
+                number_of_test_passed += 1
+            else:
+                number_of_test_failed += 1
+        print("Number of Mutants Passed: ",number_of_test_passed)
+        print("Number of Mutants Failed: ",number_of_test_failed)
+        print("Mutation Score: ", (number_of_test_failed/number_of_mutants)*100)
+
+        if not config.show_mutants:
+        # Remove mutated files and modified test files
+            for mutant_filename in mutants:
+                os.remove(mutant_filename)
+                modified_test_filename = f"{test_filename}-mutant"
+                if os.path.exists(modified_test_filename):
+                    os.remove(modified_test_filename)
+        for mutant_test_filename in test_mutants:
+            os.remove(mutant_test_filename)
+            modified_test_filename = f"{test_filename}-mutant-test"
+            if os.path.exists(modified_test_filename):
+                os.remove(modified_test_filename)
         #  Remove the folder for mutated files
         # basename = os.path.basename(original_filename)
         # dirname = os.path.dirname(original_filename)
@@ -196,6 +226,7 @@ def run_bugbane():
         # shutil.rmtree(mutant_folder)
 
 if __name__ == '__main__':
-    run_bugbane()
+    parser = build_parser()
+    run_bugbane(parser)
